Service/controller structure for basic queue operations
