+++ core/src/Application.ts
@@ -1,11 +1,13 @@
 import {ClientController} from './controller/ClientController';
 import {Logger} from './util/Logger';
 import dotenv from 'dotenv';
-import {Server} from './controller/Server';
+import {Server} from './core/Server';
+import {QueueService} from './service/QueueService';
 
 export class Application {
     private readonly rpcServer: Server;
     private clientController: ClientController;
+    private queueService: QueueService;
 
     constructor() {
         this.rpcServer = new Server({
@@ -17,15 +19,20 @@ export class Application {
     public async run(): Promise<void> {
         await this.rpcServer.start();
         // bind controllers
-        this.clientController = new ClientController(
+
+        this.queueService = new QueueService(
             {
                 messageMaxSendAttempts: parseInt(process.env.MAX_SEND_ATTEMPTS),
                 messageBaseRetryDelay: parseInt(process.env.BASE_RETRY_DELAY_MS),
-            },
+            }
+        );
+
+        this.clientController = new ClientController(
+            this.queueService,
             this.rpcServer,
         );
 
-        await this.clientController.processMessages();
+        await this.queueService.processMessages();
     }
 }
